+++ server/src/analysis.rs
